requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
PyQt5==5.15.9
requests==2.31.0
pytest==7.4.2
python-dotenv==1.0.0
SQLAlchemy==2.0.20 